+++ pxr/imaging/hd/containerDataSourceEditor.h
@@ -40,10 +40,22 @@ public:
         HdContainerDataSourceHandle initialContainer)
     : _initialContainer(initialContainer) {}
 
+    // Replaces data source at given locator and descending locations
+    // (if given a container data source) by given data source.
+    HD_API
     HdContainerDataSourceEditor &Set(
         const HdDataSourceLocator &locator,
         const HdDataSourceBaseHandle &dataSource);
 
+    // Overlays data source at given location by given data source so that
+    // data sources in the initial container at descending locations can
+    // still come through.
+    HD_API
+    HdContainerDataSourceEditor &Overlay(
+        const HdDataSourceLocator &locator,
+        const HdContainerDataSourceHandle &containerDataSource);
+
+    HD_API
     HdContainerDataSourceHandle Finish();
 
 private: