+++ pxr/usd/sdf/layerRegistry.cpp
@@ -182,26 +182,6 @@ Sdf_LayerRegistry::Find(
     } else {
         ArResolver& resolver = ArGetResolver();
 
-#if AR_VERSION == 1
-        const string layerPath = 
-            resolver.ComputeNormalizedPath(inputLayerPath);
-
-        // If the layer path is relative, this may be either a search path
-        // or a layer relative to the current working directory. Use the
-        // look-here-first scheme to check whether the registry holds a
-        // layer with the correct absolute identifier.
-        //
-        // We call TfNormPath() so we get a platform independent
-        // representation;  specifically on Windows we get forward slashes.
-        const bool isRelativePath = resolver.IsRelativePath(layerPath);
-        if (isRelativePath)
-            foundLayer = FindByIdentifier(TfNormPath(TfAbsPath(layerPath)));
-
-        // If the layer path is not relative, and we haven't found a layer
-        // yet, look up the layer using the normalized identifier.
-        if (!foundLayer && !isRelativePath)
-            foundLayer = FindByIdentifier(layerPath);
-#else
         const string& layerPath = inputLayerPath;
 
         // If the layer path depends on context there may be multiple
@@ -212,16 +192,11 @@ Sdf_LayerRegistry::Find(
         if (!resolver.IsContextDependentPath(assetPath)) {
             foundLayer = FindByIdentifier(layerPath);
         }
-#endif
 
         // If the layer path is in repository form and we haven't yet
         // found the layer via the identifier, attempt to look up the
         // layer by repository path.
-#if AR_VERSION == 1
-        const bool isRepositoryPath = resolver.IsRepositoryPath(layerPath);
-#else
         const bool isRepositoryPath = resolver.IsRepositoryPath(assetPath);
-#endif
         if (!foundLayer && isRepositoryPath)
             foundLayer = FindByRepositoryPath(layerPath);
 
@@ -327,13 +302,6 @@ Sdf_LayerRegistry::FindByRealPath(
         }
     }
     searchPath = Sdf_CreateIdentifier(searchPath, arguments);
-
-#if AR_VERSION == 1
-    // Avoid ambiguity by converting the path to a platform dependent
-    // path.  (On Windows this converts slashes to backslashes.)  The
-    // real paths stored in the registry are in platform dependent form.
-    searchPath = TfAbsPath(searchPath);
-#endif
 
     const _LayersByRealPath& byRealPath = _layers.get<by_real_path>();
     _LayersByRealPath::const_iterator realPathIt =