+++ pxr/usdImaging/usdImaging/cylinderAdapter.cpp
@@ -42,6 +42,9 @@
 
 PXR_NAMESPACE_OPEN_SCOPE
 
+namespace {
+using _PrimSource = UsdImagingDataSourceImplicitsPrim<UsdGeomCylinder, HdCylinderSchema>;
+}
 
 TF_REGISTRY_FUNCTION(TfType)
 {
@@ -76,13 +79,24 @@ UsdImagingCylinderAdapter::GetImagingSubprimData(
         const UsdImagingDataSourceStageGlobals &stageGlobals)
 {
     if (subprim.IsEmpty()) {
-        return UsdImagingDataSourceImplicitsPrim<UsdGeomCylinder>::New(
+        return _PrimSource::New(
             prim.GetPath(),
-            HdCylinderSchemaTokens->cylinder,
             prim,
             stageGlobals);
     }
     return nullptr;
+}
+
+HdDataSourceLocatorSet
+UsdImagingCylinderAdapter::InvalidateImagingSubprim(
+        TfToken const& subprim,
+        TfTokenVector const& properties)
+{
+    if (subprim.IsEmpty()) {
+        return _PrimSource::Invalidate(subprim, properties);
+    }
+    
+    return HdDataSourceLocatorSet();
 }
 
 bool