+++ pxr/usdImaging/usdImaging/stageSceneIndex.cpp
@@ -94,18 +94,23 @@ UsdImagingStageSceneIndex::_GetImagingSubprimData(
     return nullptr;
 }
 
-void
-UsdImagingStageSceneIndex::_PopulateAdapterMap()
+HdDataSourceLocatorSet
+UsdImagingStageSceneIndex::_InvalidateImagingSubprim(
+        const UsdImagingPrimAdapterSharedPtr &adapter,
+        TfToken const& subprim, TfTokenVector const& properties) const
 {
-    TRACE_FUNCTION();
-
-    UsdImagingAdapterRegistry& reg = UsdImagingAdapterRegistry::GetInstance();
-    const TfTokenVector& adapterKeys = reg.GetAdapterKeys();
-
-    for (TfToken const& adapterKey : adapterKeys) {
-        _adapterMap.insert({adapterKey, reg.ConstructAdapter(adapterKey)});
+    if (adapter) {
+        return adapter->InvalidateImagingSubprim(subprim, properties);
     }
+
+    if (subprim.IsEmpty()) {
+        return UsdImagingDataSourcePrim::Invalidate(subprim, properties);
+    }
+
+    return HdDataSourceLocatorSet();
 }
+
+
 
 UsdImagingPrimAdapterSharedPtr
 UsdImagingStageSceneIndex::_AdapterLookup(UsdPrim prim) const
@@ -116,24 +121,40 @@ UsdImagingStageSceneIndex::_AdapterLookup(UsdPrim prim) const
     // or something, instead of hardcoding the LightAPI reference...
 
     const UsdPrimTypeInfo &typeInfo = prim.GetPrimTypeInfo();
+    const TfToken adapterKey = typeInfo.GetSchemaTypeName();
 
-    _AdapterMap::const_iterator it =
-        _adapterMap.find(typeInfo.GetSchemaTypeName());
-    if(it != _adapterMap.end() && TF_VERIFY(it->second)) {
-        return it->second;
+    // If there is an adapter for the type name, return it.
+    if (UsdImagingPrimAdapterSharedPtr adapter = _AdapterLookup(adapterKey)) {
+        return adapter;
     }
 
     // XXX: Note that we're hardcoding handling for LightAPI here to match
     // UsdImagingDelegate, but the hope is to more generally support imaging
     // behaviors for API classes in the future.
+
+    // Otherwise, check for the LightAPI.
     if (prim.HasAPI<UsdLuxLightAPI>()) {
-        it = _adapterMap.find(TfToken("LightAPI"));
-        if (it != _adapterMap.end() && TF_VERIFY(it->second)) {
-            return it->second;
-        }
+        static const TfToken lightApiKey("LightAPI");
+        return _AdapterLookup(lightApiKey);
     }
 
     return nullptr;
+}
+
+UsdImagingPrimAdapterSharedPtr
+UsdImagingStageSceneIndex::_AdapterLookup(const TfToken &adapterKey) const
+{
+    // Look-up adapter in cache.
+    _AdapterMap::const_iterator const it = _adapterMap.find(adapterKey);
+    if (it != _adapterMap.end()) {
+        return it->second;
+    }
+
+    // Construct and store in cache if not in cache yet.
+    UsdImagingAdapterRegistry &reg = UsdImagingAdapterRegistry::GetInstance();
+    UsdImagingPrimAdapterSharedPtr adapter = reg.ConstructAdapter(adapterKey);
+    _adapterMap[adapterKey] = adapter;
+    return adapter;
 }
 
 // ---------------------------------------------------------------------------
@@ -264,10 +285,17 @@ void UsdImagingStageSceneIndex::SetStage(UsdStageRefPtr stage)
         TF_DEBUG(USDIMAGING_POPULATION).Msg("[Population] Removing </>\n");
         _SendPrimsRemoved({SdfPath::AbsoluteRootPath()});
         _stageGlobals.Clear();
+        TfNotice::Revoke(_objectsChangedNoticeKey);
         _adapterMap.clear();
     }
+
     _stage = stage;
-    _PopulateAdapterMap();
+
+    if (_stage) {
+        _objectsChangedNoticeKey =
+            TfNotice::Register(TfCreateWeakPtr(this),
+                &UsdImagingStageSceneIndex::_OnUsdObjectsChanged, _stage);
+    }
 }
 
 void UsdImagingStageSceneIndex::Populate()
@@ -343,6 +371,153 @@ UsdImagingStageSceneIndex::_GetTraversalPredicate() const
     // UsdPrimRange traversal isn't impossible to follow.  For now we'll go
     // with the default predicate, and resolve special cases as they come up.
     return UsdPrimDefaultPredicate;
+}
+
+// ---------------------------------------------------------------------------
+
+void
+UsdImagingStageSceneIndex::_OnUsdObjectsChanged(
+    UsdNotice::ObjectsChanged const& notice,
+    UsdStageWeakPtr const& sender)
+{
+    if (!sender || !TF_VERIFY(sender == _stage)) {
+        return;
+    }
+
+    TRACE_FUNCTION();
+
+    TF_DEBUG(USDIMAGING_CHANGES).Msg("[Objects Changed] Notice received "
+            "from stage with root layer @%s@\n",
+            sender->GetRootLayer()->GetIdentifier().c_str());
+
+    // These paths represent objects which have been modified in a structural
+    // way, for example changing type or composition topology. These paths may
+    // be paths to prims or properties. Prim resyncs trigger a repopulation of
+    // the subtree rooted at the prim path. Property resyncs are promoted to
+    // hydra property invalidations.
+    const UsdNotice::ObjectsChanged::PathRange pathsToResync =
+        notice.GetResyncedPaths();
+    for (auto it = pathsToResync.begin(); it != pathsToResync.end(); ++it) {
+        if (it->IsPrimPath()) {
+            _usdPrimsToResync.push_back(*it);
+            TF_DEBUG(USDIMAGING_CHANGES).Msg(" - Resync queued: %s\n",
+                    it->GetText());
+        } else if (it->IsPropertyPath()) {
+            _usdPropertiesToUpdate[it->GetPrimPath()]
+                .push_back(it->GetNameToken());
+            TF_DEBUG(USDIMAGING_CHANGES).Msg(
+                    " - Property update due to property resync queued: %s\n",
+                    it->GetText());
+        }
+    }
+
+    // These paths represent objects which have been modified in a 
+    // non-structural way, for example setting a value. These paths may be paths
+    // to prims or properties. Property invalidations flow into hydra as dirty
+    // locators. Prim invalidations are promoted to resyncs or ignored.
+    const UsdNotice::ObjectsChanged::PathRange pathsToUpdate =
+        notice.GetChangedInfoOnlyPaths();
+    const SdfSchema& schema = SdfSchema::GetInstance();
+
+    for (auto it = pathsToUpdate.begin(); it != pathsToUpdate.end(); ++it) {
+        if (it->IsPrimPath()) {
+            // By default, resync the prim if there are any changes to plugin
+            // fields and ignore changes to built-in fields. Schemas typically
+            // register their own plugin metadata fields instead of relying on
+            // built-in fields.
+            const TfTokenVector changedFields = it.GetChangedFields();
+            for (const TfToken &field : changedFields) {
+                const SdfSchema::FieldDefinition *fieldDef =
+                    schema.GetFieldDefinition(field);
+                if (fieldDef && fieldDef->IsPlugin()) {
+                    _usdPrimsToResync.push_back(*it);
+                    TF_DEBUG(USDIMAGING_CHANGES).Msg(
+                            " - Resync due to prim update queued: %s\n",
+                            it->GetText());
+                    break;
+                }
+            }
+        } else if (it->IsPropertyPath()) {
+            _usdPropertiesToUpdate[it->GetPrimPath()]
+                .push_back(it->GetNameToken());
+            TF_DEBUG(USDIMAGING_CHANGES).Msg(" - Property update queued: %s\n",
+                    it->GetText());
+        }
+    }
+}
+
+void
+UsdImagingStageSceneIndex::ApplyPendingUpdates()
+{
+    if (!_stage ||
+        (_usdPrimsToResync.empty() && _usdPropertiesToUpdate.empty())) {
+        return;
+    }
+
+    TRACE_FUNCTION();
+
+    // Resync first...
+    std::sort(_usdPrimsToResync.begin(), _usdPrimsToResync.end());
+    size_t lastResynced = 0;
+    for (size_t i = 0; i < _usdPrimsToResync.size(); ++i) {
+        // Coalesce paths with a common prefix, so as not to resync /A and /A/B,
+        // since due to their hierarchical nature the latter is redundant.
+        // Thanks to the sort, all suffixes of path[i] are in a contiguous block
+        // to the right of i.  We skip all resync paths until we find one that's
+        // not a suffix of path[i], which marks the start of a new (possibly
+        // 1-element) contiguous block of suffixes of some path.
+        if (i > 0 && _usdPrimsToResync[i].HasPrefix(
+                _usdPrimsToResync[lastResynced])) {
+            continue;
+        }
+        lastResynced = i;
+
+        TF_DEBUG(USDIMAGING_POPULATION).Msg("[Population] Removing <%s>\n",
+                _usdPrimsToResync[i].GetText());
+        _SendPrimsRemoved({_usdPrimsToResync[i]});
+        _Populate(_stage->GetPrimAtPath(_usdPrimsToResync[i]));
+
+        // Prune property updates of resynced prims, which are redundant.
+        auto start = _usdPropertiesToUpdate.lower_bound(_usdPrimsToResync[i]);
+        auto end = start;
+        while (end != _usdPropertiesToUpdate.end() &&
+               end->first.HasPrefix(_usdPrimsToResync[i])) {
+            ++end;
+        }
+        if (start != end) {
+            _usdPropertiesToUpdate.erase(start, end);
+        }
+    }
+
+    // Changed properties...
+    HdSceneIndexObserver::DirtiedPrimEntries dirtiedPrims;
+    for (auto const& pair : _usdPropertiesToUpdate) {
+        const SdfPath &primPath = pair.first;
+        const TfTokenVector &properties = pair.second;
+        // XXX: We could sort/unique the properties here...
+        
+        const UsdPrim prim = _stage->GetPrimAtPath(primPath);
+        UsdImagingPrimAdapterSharedPtr adapter =
+            _AdapterLookup(prim);
+
+        const TfTokenVector subprims = _GetImagingSubprims(adapter);
+        for (TfToken const& subprim : subprims) {
+            HdDataSourceLocatorSet dirtyLocators =
+                _InvalidateImagingSubprim(adapter, subprim, properties);
+            if (!dirtyLocators.IsEmpty()) {
+                SdfPath const subpath = subprim.IsEmpty()
+                    ? primPath : primPath.AppendChild(subprim);
+                dirtiedPrims.emplace_back(subpath, dirtyLocators);
+            }
+        }
+    }
+
+    _usdPrimsToResync.clear();
+    _usdPropertiesToUpdate.clear();
+
+    if (dirtiedPrims.size() > 0) {
+        _SendPrimsDirtied(dirtiedPrims);
+    }
 }
 
 // ---------------------------------------------------------------------------